requirements.txt
leancloud-sdk
flask
numpy
scikit-learn
blinker
logentries
bugsnag